Split into multiple modules
This splits the code into various modules for
a cleaner interface and code readability.

There is now an interface for date operations
and conversions which abstracts the conversion
type being performed.
